ML/Percobaan/requirements.txt
keras_facenet==0.3.2
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==10.3.0